+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased](https://github.com/nuwave/lighthouse/compare/v4.1.1...master)
 
+### Fixed
+
+- The connection type was not based on the edge type when the edge type was specified for the relay type https://github.com/nuwave/lighthouse/pull/939
+
 ## [4.1.1](https://github.com/nuwave/lighthouse/compare/v4.1.0...v4.1.1)
 
 ### Fixed