+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased](https://github.com/nuwave/lighthouse/compare/v4.2.0...master)
 
+### Fixed
+
+- Actually use the specified `edgeType` in Relay style connections https://github.com/nuwave/lighthouse/pull/939
+
 ## [4.2.0](https://github.com/nuwave/lighthouse/compare/v4.1.1...v4.2.0)
 
 ### Added